Add a function for decoding physical signals (#14)
EDF stores its signals encoded as two-byte integers (`Int16`). We can
perform a simple reverse transformation to recover the physical signal.
The package now exposes that as a `decode` function that returns a
vector when given a `Signal`. The formula used for the reverse
transformation is the same as in Pixpipe's JavaScript EDF reader.