Remove `nb_header` from the `EDF.Header` specification
The value generated by nb_header can be calculated pretty easily during the process of
writing to EDF. Removing this field allows the caller of `EDF.Header` to create an instance
without needing to specify the number of bytes in the header in advance.

The test for `nb_header` has additionally been rewritten to test that the value calculated
in the revised version of `EDF.write` is correct.