Add `SignalHeader` struct
This is part of an incremental refactor in which metadata regarding various
signals is separated from the Signal struct itself. This will allow for
future write methods to dispatch on components of the signals, rather than
in batches on the entire signal. This will reduce code complexity while
retaining the fundamental structure of EDF.File itself, and will make future
Mmapping endeavors a bit easier.

Changes to read, write, and runtests are purely to maintain passing tests
while these incremental changes occur.