Rewrite annotation structs
The first annotation in an EDF data record was originally the fields in a
`RecordAnnotation` instance, along with an optional list of additional
`TimestampAnnotationList` instances. Since both are roughly written the same
way, with minor differences, this has been refactored into two subtypes of
`AbstractAnnotation` that, in the future, should be able to share a `write`
method.

The convenience aliases `DataRecord` and `TimestampAnnotationList` are used
for cleaner representations of field types in the structs that use them.

`read` and `write` have received minor touch-ups to work with the new structs.
Additionally, no information about byte count is required by anyone who
constructs any kind of annotation-related struct anymore.

Tests have been updated to work with the new structs, and docs are updated.