Fiddle with CI a bit (#63)
* Fiddle with CI a bit

Codecov didn't report a status in a recent PR even though their website
shows they received the coverage information. Codecov doesn't require a
token to be passed for public repositories but we're passing one anyway
so I'm deleting it and seeing what happens. While I'm touching that
file, I also simplified the artifact caching to use a Julia-specific
GitHub Action that handles all of that. Lastly, I deleted the Codecov
YAML file that just tells it not to comment. Comments are fine, who
cares.

* Try putting back the token? I gave it a real one

* Bump `codecov-action` to v3

* Stuff